% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{>=,cspec,numeric-method}
\alias{>=,cspec,numeric-method}
\alias{>,cspec,numeric-method}
\alias{<=,cspec,numeric-method}
\alias{<,cspec,numeric-method}
\alias{==,cspec,numeric-method}
\alias{>=,numeric,cspec-method}
\alias{>,numeric,cspec-method}
\alias{<=,numeric,cspec-method}
\alias{<,numeric,cspec-method}
\alias{==,numeric,cspec-method}
\title{Comparison methods for content specifications}
\usage{
\S4method{>=}{cspec,numeric}(e1, e2)

\S4method{>}{cspec,numeric}(e1, e2)

\S4method{<=}{cspec,numeric}(e1, e2)

\S4method{<}{cspec,numeric}(e1, e2)

\S4method{==}{cspec,numeric}(e1, e2)

\S4method{>=}{numeric,cspec}(e1, e2)

\S4method{>}{numeric,cspec}(e1, e2)

\S4method{<=}{numeric,cspec}(e1, e2)

\S4method{<}{numeric,cspec}(e1, e2)

\S4method{==}{numeric,cspec}(e1, e2)
}
\arguments{
\item{e1}{content specification or a scalar number}

\item{e2}{content specification or a scalar number}
}
\description{
Comparison methods for content specifications
}
