% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{difficulty_constraint}
\alias{difficulty_constraint}
\title{Constraints on test or item difficulty}
\usage{
difficulty_constraint(
  avg_pval_min = NULL,
  avg_pval_max = NULL,
  pval_min = NULL,
  pval_max = NULL,
  modules = NULL
)
}
\arguments{
\item{avg_pval_min}{minimum allowed average pvalue (proportion correct) on the test}

\item{avg_pval_max}{maximum allowed average pvalue on the test}

\item{pval_min}{minimum allowed pvalue for any item in the test}

\item{pval_max}{maximum allowed pvalue for any item in the test}

\item{modules}{integer vector of modules to apply to, with the 'easiest' module specified by 1. If NULL, applies to all modules}
}
\description{
These contraints use the item parameters and population distribution to limit the difficulty of the test or individual items
to a specified range
}
\details{
You must specify at least one argument.
}
