% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{irt_constraint}
\alias{irt_constraint}
\title{IRT constraints on test designs}
\usage{
irt_constraint(theta, min_information = NULL, max_sem = NULL)
}
\arguments{
\item{theta}{points on the ability scale to evaluate information}

\item{min_information}{the minimum information required}

\item{max_sem}{the maximum allowed Standard Error of Measurement}
}
\description{
Typically used when the length of the test is minimized. You must specify either min_information or max_sem but not both.
}
