% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{item_fraction}
\alias{item_fraction}
\alias{item_sum}
\alias{item_count}
\title{Content specification}
\usage{
item_fraction(predicate, subset = NULL, paths = NULL)

item_sum(predicate, paths = NULL)

item_count(predicate = NULL, paths = NULL)
}
\arguments{
\item{predicate}{predicate}

\item{subset}{subset}

\item{paths}{paths to which the specification should apply. NULL means all paths. Paths are indicated by consecutive integers
where 1 stands for the path to which persons with the lowest score on the routing tests are routed. Argument is ignored for linear tests}
}
\description{
Specify content for test designs. These functions are meant to be used with \code{\link{test_design}} or \code{\link{mst_design}}
}
