% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{minimize_items}
\alias{minimize_items}
\alias{maximize_items}
\alias{random_test}
\alias{maximize_information}
\title{Property of the test to optimize}
\usage{
minimize_items(property = NULL)

maximize_items(property = NULL)

random_test()

maximize_information(theta = 0, weight = 1)
}
\arguments{
\item{property}{unquoted column name or statement to be evaluated in the context of items, e.g. item_max_score, time_required. If NULL
optimizes the number of items in the test.}

\item{theta}{point(s) on the ability scale on which the information should be maximized}

\item{weight}{relative weights attached to the information on theta}
}
\description{
Minimize_items or maximize_items minimize of maximize some property of the items in the test. 
Random_test creates a random selection of items.
Maximize_information maximizes the information on specified points of the ability scale.
}
