% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mst_design}
\alias{plot_mst_design}
\title{IRT plot of mst design effects}
\usage{
plot_mst_design(
  md,
  pars,
  populaton_density = dnorm,
  type = c("sem", "info"),
  infolines = NULL
)
}
\arguments{
\item{md}{mst design, output of function \code{\link{mst_design}}}

\item{pars}{item parameters}

\item{populaton_density}{density function for the population, the default is standard normal.}

\item{type}{information or SEM}

\item{infolines}{for the information plot vertical lines at these points}
}
\value{
ggplot2 like object to be plotted
}
\description{
IRT plot of mst design effects
}
