% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{pretest_design}
\alias{pretest_design}
\title{Simple pretest designs}
\usage{
pretest_design(items, max_nit, balance = NULL, friends = NULL)
}
\arguments{
\item{items}{data.frame with columns item_id and optionally weight and other columns. 
See details}

\item{max_nit}{maximum number of items in a booklet; if items contains a column
`weight` then this is the maximum sum of item weight (e.g. expected time spent on an item)}

\item{balance}{categorical variables in items that will be balanced across booklets}

\item{friends}{name of an id column in items that specifies friend groups (e.g. reading texts)}
}
\value{
data.frame
}
\description{
Simple pretest designs
}
\details{
This function returns a simple chained block design which minimizes the number of booklet versions

to do: items$weight does not yet do anything
}
