% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply.R
\name{size_groups}
\alias{size_groups}
\title{Create a grouping variable by size}
\usage{
size_groups(size, k = NULL, n = NULL)
}
\arguments{
\item{size}{typically an integer (e.g. number of persons within schools)}

\item{k}{number of groups (groupsize is variable)}

\item{n}{maximum size of groups (number of groups is variable)}
}
\description{
divide objects into bins of equal size. Uses a sort of greedy method
}
\details{
You must specify only one of k or n. Since bin packing and the multi-way partition problem are NP-hard,
the solution is not guaranteed to be optimal. 

Bin packing (i.e. maximum size groups), puts items of larger size together, 
the number of items in each bin thus tends to be very unequal
}
