% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd.R
\name{pretest_design}
\alias{pretest_design}
\title{Simple pretest designs}
\usage{
pretest_design(items, max_nit, balance = NULL, friends = NULL)
}
\arguments{
\item{items}{data.frame with columns item_id and variables listed under "balance" or "friends"}

\item{max_nit}{maximum number of items in a booklet.}

\item{balance}{categorical variables in items that will be balanced across booklets}

\item{friends}{name of an id column in items that specifies friend groups (e.g. reading texts)}
}
\value{
data.frame
}
\description{
Simple pretest designs
}
\details{
This function returns a simple chained block design which minimizes the number of booklet versions
}
